requirements.txt
flask
sentence-transformers
faiss-cpu
transformers
torch
openai
numpy
tqdm
gunicorn
sounddevice
scipy
openai-whisper
pyttsx3
